\name{BattingPlayingTime2014}
\docType{data}
\alias{BattingPlayingTime2014}
\title{2014 Playing Time Projections for Hitters}
\description{
  This data gives playing time projections for MLB hitters.
}
\usage{BattingPlayingTime2014}
\format{A data.frame with 530 rows and 3 columns.}
\keyword{datasets}
