\name{lastplayerpicked}
\alias{lastplayerpicked-package}
\alias{lastplayerpicked}
\docType{package}
\title{Last Player Picked Price Guide}
\description{
Generates customized fantasy baseball dollar values.

This code is a re-write of the now defunct Last Player Picked (LPP) web site.
Thanks and credit should should go to that author (pseudonym Mays Copeland),
who made the source code available at \url{https://github.com/mayscopeland/priceguide}.

The 2014 Steamer, Oliver, and ZiPS projections were downloaded from \url{http://www.fangraphs.com}.
2014 playing time estimations came from \url{http://www.fantasypros.com/mlb/projections/hitters.php}.
2013 position eligibility data came from \url{http://www.rotochamp.com/}.
Current MLB IDs came from \url{http://www.baseballprospectus.com/sortable/playerid_list.php}.
Some data wrangling was necessary to convert these files to a LPP-friendly version.

The real Mays Copeland was a pitcher in Major League Baseball.
He played one game for the St. Louis Cardinals in 1935.
}

\author{
Cole Beck \email{cole.beck@gmail.com}
}
