\name{PitchingPlayingTime2014}
\docType{data}
\alias{PitchingPlayingTime2014}
\title{2014 Playing Time Projections for Pitchers}
\description{
  This data gives playing time projections for MLB pitchers.
}
\usage{PitchingPlayingTime2014}
\format{A data.frame with 457 rows and 4 columns.}
\keyword{datasets}
